fix: flag is not erroneously double counted or attributed to the wrong side
Due to issues in how resetFlag() was handled, the flag would sometimes be counted
multiple times for a singular capture or also be attributed to a side that should
not have received credit. Some refactoring changes were incorporated in this too;
see the diffs for details.

I will be pushing a refactoring commit after this.